+++ js/util.js
@@ -1,0 +1,10 @@
+function getRandomInteger(min, max) {
+  min = Math.ceil(min);
+  max = Math.floor(max);
+  if (min < 0 || max <= min) {
+    return NaN;
+  }
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
+
+export {getRandomInteger};
